extract homescreen reducer into flux/parts/homescreen
fix import for updateViewOrder test
